Add vanilla CSS support to Select (#519)
* Add vanilla CSS support to Select

* PR feedback

* Don't wrap portal + content
