+++ components/demos/index.jsx
@@ -15,7 +15,7 @@ export { default as PopoverDemo } from './Popover';
 export { default as ProgressDemo } from './Progress';
 export { default as RadioGroupDemo } from './RadioGroup';
 export { default as ScrollAreaDemo } from './ScrollArea';
-export { default as SelectDemo } from './Select';
+export { default as SelectDemo } from './Select/stitches';
 export { default as SeparatorDemo } from './Separator/stitches';
 export { default as SliderDemo } from './Slider/stitches';
 export { default as SwitchDemo } from './Switch/stitches';
